switch free_page() from unsigned long to const void *

+++ arch/alpha/include/asm/pgalloc.h
@@ -34,7 +34,7 @@ extern pgd_t *pgd_alloc(struct mm_struct *mm);
 static inline void
 pgd_free(struct mm_struct *mm, pgd_t *pgd)
 {
-	free_page((unsigned long)pgd);
+	free_page(pgd);
 }
 
 static inline pmd_t *
@@ -47,7 +47,7 @@ pmd_alloc_one(struct mm_struct *mm, unsigned long address)
 static inline void
 pmd_free(struct mm_struct *mm, pmd_t *pmd)
 {
-	free_page((unsigned long)pmd);
+	free_page(pmd);
 }
 
 static inline pte_t *
@@ -60,7 +60,7 @@ pte_alloc_one_kernel(struct mm_struct *mm, unsigned long address)
 static inline void
 pte_free_kernel(struct mm_struct *mm, pte_t *pte)
 {
-	free_page((unsigned long)pte);
+	free_page(pte);
 }
 
 static inline pgtable_t
