+++ arch/hexagon/include/asm/pgalloc.h
@@ -56,7 +56,7 @@ static inline pgd_t *pgd_alloc(struct mm_struct *mm)
 
 static inline void pgd_free(struct mm_struct *mm, pgd_t *pgd)
 {
-	free_page((unsigned long) pgd);
+	free_page(pgd);
 }
 
 static inline struct page *pte_alloc_one(struct mm_struct *mm,
@@ -90,7 +90,7 @@ static inline void pte_free(struct mm_struct *mm, struct page *pte)
 
 static inline void pte_free_kernel(struct mm_struct *mm, pte_t *pte)
 {
-	free_page((unsigned long)pte);
+	free_page(pte);
 }
 
 static inline void pmd_populate(struct mm_struct *mm, pmd_t *pmd,
