+++ arch/m32r/include/asm/pgalloc.h
@@ -27,7 +27,7 @@ static __inline__ pgd_t *pgd_alloc(struct mm_struct *mm)
 
 static inline void pgd_free(struct mm_struct *mm, pgd_t *pgd)
 {
-	free_page((unsigned long)pgd);
+	free_page(pgd);
 }
 
 static __inline__ pte_t *pte_alloc_one_kernel(struct mm_struct *mm,
@@ -54,7 +54,7 @@ static __inline__ pgtable_t pte_alloc_one(struct mm_struct *mm,
 
 static inline void pte_free_kernel(struct mm_struct *mm, pte_t *pte)
 {
-	free_page((unsigned long)pte);
+	free_page(pte);
 }
 
 static inline void pte_free(struct mm_struct *mm, pgtable_t pte)
