+++ arch/m68k/include/asm/sun3_pgalloc.h
@@ -19,7 +19,7 @@ extern const char bad_pmd_string[];
 
 static inline void pte_free_kernel(struct mm_struct *mm, pte_t *pte)
 {
-        free_page((unsigned long) pte);
+        free_page(pte);
 }
 
 static inline void pte_free(struct mm_struct *mm, pgtable_t page)
@@ -83,7 +83,7 @@ static inline void pmd_populate(struct mm_struct *mm, pmd_t *pmd, pgtable_t page
 
 static inline void pgd_free(struct mm_struct *mm, pgd_t *pgd)
 {
-        free_page((unsigned long) pgd);
+        free_page(pgd);
 }
 
 static inline pgd_t * pgd_alloc(struct mm_struct *mm)
